+++ mmdet3d/ops/__init__.py
@@ -12,7 +12,8 @@ from .interpolate import three_interpolate, three_nn
 from .knn import knn
 from .norm import NaiveSyncBatchNorm1d, NaiveSyncBatchNorm2d
 from .paconv import PAConv, PAConvCUDA, assign_score_withk
-from .pointnet_modules import (PAConvSAModule, PAConvSAModuleMSG,
+from .pointnet_modules import (PAConvSAModule, PAConvSAModuleCUDA,
+                               PAConvSAModuleMSG, PAConvSAModuleMSGCUDA,
                                PointFPModule, PointSAModule, PointSAModuleMSG,
                                build_sa_module)
 from .roiaware_pool3d import (RoIAwarePool3d, points_in_boxes_batch,
@@ -34,5 +35,6 @@ __all__ = [
     'QueryAndGroup', 'PointSAModule', 'PointSAModuleMSG', 'PointFPModule',
     'points_in_boxes_batch', 'get_compiler_version', 'assign_score_withk',
     'get_compiling_cuda_version', 'Points_Sampler', 'build_sa_module',
-    'PAConv', 'PAConvCUDA', 'PAConvSAModuleMSG', 'PAConvSAModule'
+    'PAConv', 'PAConvCUDA', 'PAConvSAModuleMSG', 'PAConvSAModule',
+    'PAConvSAModuleCUDA', 'PAConvSAModuleMSGCUDA'
 ]
