+++ mmdet3d/ops/pointnet_modules/__init__.py
@@ -1,9 +1,11 @@
 from .builder import build_sa_module
 from .point_fp_module import PointFPModule
-from .point_sa_module import (PAConvSAModule, PAConvSAModuleMSG, PointSAModule,
-                              PointSAModuleMSG)
+from .point_sa_module import (PAConvSAModule, PAConvSAModuleCUDA,
+                              PAConvSAModuleMSG, PAConvSAModuleMSGCUDA,
+                              PointSAModule, PointSAModuleMSG)
 
 __all__ = [
     'build_sa_module', 'PointSAModuleMSG', 'PointSAModule', 'PointFPModule',
-    'PAConvSAModule', 'PAConvSAModuleMSG'
+    'PAConvSAModule', 'PAConvSAModuleMSG', 'PAConvSAModuleCUDA',
+    'PAConvSAModuleMSGCUDA'
 ]
